work with some warnings under windows

+++ pkg/js/include/Boom/JS/Scope.hpp
@@ -14,17 +14,17 @@ public:
     Scope(boom::js::ContextRef, void*, void**, std::size_t);
     boom::js::ContextRef context() const;
     boom::js::ValueRef thisObject() const;
-    std::int32_t argCount() const;
-    boom::js::ValueRef getArg(std::int32_t) const;
+    std::size_t argCount() const;
+    boom::js::ValueRef getArg(std::size_t) const;
     virtual ~Scope();
 private:
     boom::js::ContextRef _context;
-    std::int32_t _argc;
+    std::size_t _argc;
     boom::js::__ScopeImpl* _impl;
     void _implInit(void*, void**, std::size_t);
     void _implDone();
     boom::js::ValueRef _implThisObject() const;
-    boom::js::ValueRef _implGetArg(std::int32_t) const;
+    boom::js::ValueRef _implGetArg(std::size_t) const;
 };
 
 } /* namespace boom::js */
